+++ core/components/ValidationError.js
@@ -1,9 +1,0 @@
-export default {
-  name: 'ValidationError',
-  props: {
-    message: {
-      type: String,
-      default: ''
-    }
-  }
-}