+++ src/themes/default/components/core/ColorSelector.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import GenericSelector from '@vue-storefront/core/components/GenericSelector'
+import GenericSelector from '@vue-storefront/core/compatibility/components/GenericSelector'
 
 export default {
   mixins: [GenericSelector],