+++ src/themes/default/components/core/blocks/Form/BaseCheckbox.vue
@@ -32,10 +32,28 @@
 </template>
 
 <script>
-import baseCheckbox from '@vue-storefront/core/components/blocks/Form/BaseCheckbox'
-
 export default {
-  mixins: [baseCheckbox]
+  name: 'BaseCheckbox',
+  props: {
+    id: {
+      type: String,
+      required: true
+    },
+    value: {
+      type: Boolean,
+      required: true
+    },
+    validation: {
+      type: Object,
+      required: false,
+      default: () => { }
+    },
+    disabled: {
+      type: Boolean,
+      required: false,
+      default: false
+    }
+  }
 }
 </script>