+++ src/themes/default/components/core/blocks/SidebarMenu/SidebarMenu.vue
@@ -117,10 +117,10 @@
 </template>
 
 <script>
-import { mapState, mapGetters } from 'vuex'
+import { mapState } from 'vuex'
 import i18n from '@vue-storefront/i18n'
 
-import SidebarMenu from '@vue-storefront/core/components/blocks/SidebarMenu/SidebarMenu'
+import SidebarMenu from '@vue-storefront/core/compatibility/components/blocks/SidebarMenu/SidebarMenu'
 import SubBtn from 'theme/components/core/blocks/SidebarMenu/SubBtn'
 import SubCategory from 'theme/components/core/blocks/SidebarMenu/SubCategory'
 
@@ -173,9 +173,6 @@ export default {
     ...mapState({
       submenu: state => state.ui.submenu,
       currentUser: state => state.user.current
-    }),
-    ...mapGetters('compare', {
-      compareIsActive: 'isActive'
     })
   },
   methods: {