+++ CHANGELOG.md
@@ -8,16 +8,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Lazy loading for SSR and non-SSR routes
-- Many modules
+- app splitted into modules
 
 ### Removed
 - `vsf-payment-stripe` module integration removed from core
 
 ### Changed
-- ProductSlider, ProductLinks, ProductListing moved to theme. Other Product-related components are now served directly from module
-- Many theme-related logic moved to theme (d+ eleted empty core components just with `name`)
+- There is new config option `config.orders.directBackendSync` that changes the behavior of placing an order. Please do read  [more on this change](https://github.com/DivanteLtd/vue-storefront/commit/e73f2ca19a5d33a39f8b0fd6346543eced24167e) and [more on vue-storefront-api change](https://github.com/DivanteLtd/vue-storefront-api/commit/80c497f72362c72983db4fdcac14c8ba6f8729a8)
+- ProductSlider, ProductLinks, ProductListing moved to theme. 
+- Many theme-related logic moved to theme (+ deleted empty core components just with `name`)
 - Components required for backward compatibility moved to `compatibility` folder. For all this files you just need to add `compatibility` after `core` in import path to make them work like before.
 - Better Vuex extensibility with modules 
+- VSModule `store` object changed to fulfil need of multiple vuex modules (see modules docs)
+- UI Store logic for Microcart moved to cart module
+- Extensions are now depreciated, theme-level extensions removed and src-level extension to be depreciated in 1.7
+- Theme-starter depreciated and removed (will be replaced with theme 2.0)
+- Header, Form components, (baseCheckbox, BaseInput, BaseRadioButton, BaseSelect, Basetextarea) Loader, MainSlider, Footer, SearchIcon, ForgotPass, SignUp and Modal core components moved to theme
+- extendStore deprecaiated and moved to compatibility folder
 
 ## [1.5.0] - 2018.10.22